\name{acvs.to.acs}
\alias{acvs.to.acs}
\alias{acvf.to.acf}
\title{Convert from the ACVS to the ACS}
\description{Converts a vector of autocovariances to a vector of autocorrelations}
\usage{ acvs.to.acs(the.acvs) }
\arguments{
\item{the.acvs}{The vector autocovariances to covert}
}
\value{The vector of autocorelations}
\examples{ }
\author{Peter F. Craigmile}
\keyword{}
