\name{AIC.to.AICC}
\alias{AIC.to.AICC}
\title{Converts from the AICC to the AIC.}
\description{Converts from the AICC to the AIC.}
\usage{AIC.to.AICC(aic, n, npars)}
\arguments{
  \item{aic}{The AIC value}
  \item{n}{The length of the time series}
  \item{npars}{The number of parameters}
}
\value{The AICC value}
\examples{ }
\author{Peter F. Craigmile}
\keyword{}
