\name{AICC}
\alias{AICC}
\title{Calculates the corrected Akaike's information criterion (AICC).}
\description{Calculates the corrected Akaike's information criterion (AICC).}
\usage{AICC(model)}
\arguments{
  \item{model}{an object of class 'Arima'}
}
\value{The value of the AICC}
\examples{ }
\author{Peter F. Craigmile}
\keyword{}
