\name{FPE}
\alias{FPE}
\title{Calculates Akaike's Final Prediction Error (FPE) criterion}
\description{Calculates Akaike's Final Prediction Error (FPE)
 criterion for the AR 'model' fitted using the 'arima' function.}
\usage{FPE(model)}
\arguments{
  \item{model}{an object of class 'Arima'}
}
\value{The value of the final prediction error}
\examples{ }
\author{Peter F. Craigmile}
\keyword{}
